Task 049: Fixed, cambiado de int a double las horas

+++ src/main/java/acme/entities/enrolment/Enrolment.java
@@ -42,5 +42,5 @@ public class Enrolment extends AbstractEntity {
 	protected String			goals;
 
 	@PositiveOrZero
-	protected int				workTime;
+	protected double			workTime;
 }
